translate: split TranslationSet to own file
Split TranslationSet code into its own file, having everything in one
file was getting messy.
